+++ mangooio-core/src/main/java/io/mangoo/monitoring/Telemetry.java
@@ -7,8 +7,9 @@ import io.opentelemetry.api.OpenTelemetry;
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.trace.Span;
+import io.opentelemetry.api.trace.SpanKind;
 import io.opentelemetry.api.trace.Tracer;
-import io.opentelemetry.context.Context;
+import io.opentelemetry.context.Scope;
 import io.opentelemetry.exporter.otlp.trace.OtlpGrpcSpanExporter;
 import io.opentelemetry.sdk.OpenTelemetrySdk;
 import io.opentelemetry.sdk.resources.Resource;
@@ -16,6 +17,8 @@ import io.opentelemetry.sdk.trace.SdkTracerProvider;
 import io.opentelemetry.sdk.trace.export.BatchSpanProcessor;
 import jakarta.inject.Inject;
 import jakarta.inject.Singleton;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import java.util.Map;
 import java.util.Objects;
@@ -23,13 +26,13 @@ import java.util.concurrent.ConcurrentHashMap;
 
 @Singleton
 public class Telemetry {
+    private static final Logger LOG = LogManager.getLogger(Telemetry.class);
     private final Map<String, Span> spans = new ConcurrentHashMap<>();
     private SdkTracerProvider traceProvider;
     private OpenTelemetry openTelemetry;
-    private Tracer tracer;
 
     @Inject
-    public Telemetry (Config config) {
+    public Telemetry(Config config) {
         Objects.requireNonNull(config, Required.CONFIG);
 
         if (config.isOtlpEnable()) {
@@ -47,8 +50,6 @@ public class Telemetry {
             this.openTelemetry = OpenTelemetrySdk.builder()
                     .setTracerProvider(this.traceProvider)
                     .build();
-
-            this.tracer = openTelemetry.getTracer(config.getOtlpScopeName());
         }
     }
 
@@ -58,19 +59,33 @@ public class Telemetry {
         }
     }
 
-    public void begin(String process) {
+    public void begin(String process, String scope) {
         Arguments.requireNonBlank(process, Required.PROCESS);
-        spans.put(process, tracer.spanBuilder(process).startSpan());
+        Arguments.requireNonBlank(scope, Required.SCOPE);
+
+        Tracer tracer = openTelemetry.getTracer(scope);
+        Span span = tracer.spanBuilder(process)
+                .setSpanKind(SpanKind.INTERNAL)
+                .startSpan();
+
+        spans.put(process, span);
     }
 
-    public void begin(String parentProcess, String childProcess) {
+    public void begin(String parentProcess, String childProcess, String scope) {
+        Arguments.requireNonBlank(parentProcess, Required.PROCESS);
+        Arguments.requireNonBlank(childProcess, Required.PROCESS);
+        Arguments.requireNonBlank(scope, Required.SCOPE);
+
         Span parentSpan = spans.get(parentProcess);
         if (parentSpan != null) {
-            Span child = tracer.spanBuilder(childProcess)
-                    .setParent(Context.current().with(parentSpan))
-                    .startSpan();
+            try (Scope ignored = parentSpan.makeCurrent()) {
+                Tracer tracer = openTelemetry.getTracer(scope);
+                Span child = tracer.spanBuilder(childProcess)
+                        .setSpanKind(SpanKind.INTERNAL)
+                        .startSpan();
 
-            spans.put(childProcess, child);
+                spans.put(childProcess, child);
+            }
         }
     }
 
@@ -79,9 +94,13 @@ public class Telemetry {
 
         Span span = spans.get(process);
         if (span != null) {
-            span.end();
+            try {
+                span.end();
+            } catch (Exception e) {
+                LOG.error("Fail to end span", e);
+            }
+            spans.remove(process);
         }
-        spans.remove(process);
     }
 
     public OpenTelemetry getOpenTelemetry() {