+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 Unreleased
 
+* Added Support for Open Telemetry (svenkubiak)
 * Added new vault functionality for storing secrets based on Java Keystore (svenkubiak)
 * Added HTTPS connector (svenkubiak)
 * Removed cryptex{} from configuration in favour of new vault (svenkubiak)